Use TabbedPanel for the help tabs
The hand-written tab bar is replaced by TabbedPanel, which already handles
switching and icons, so the feature does not carry its own copy. Chapters of
a single tab moved to the new HelpChaptersPanel, which makes HelpContentModel
unnecessary - the panel takes the tabs directly.

Task: 12033
