+++ gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/component/wizard/collapse/HelpChaptersPanel.html
@@ -1,0 +1,32 @@
+<!--
+  ~ Copyright (C) 2026 Evolveum and contributors
+  ~
+  ~ Licensed under the EUPL-1.2 or later.
+  -->
+
+<!DOCTYPE html>
+<html xmlns:wicket="http://wicket.apache.org">
+<wicket:panel>
+    <div class="d-flex flex-column h-100">
+
+        <form class="d-flex flex-row align-items-center gap-1 pb-2" wicket:id="pagerForm">
+            <a class="btn btn-sm btn-outline-secondary px-2" wicket:id="previousChapter"
+               wicket:message="title:HelpContentPanel.previousChapter">
+                <i class="fa fa-arrow-left"></i>
+            </a>
+            <span class="border rounded px-2 py-1 small text-secondary text-nowrap"
+                  wicket:id="chapterCounter"></span>
+            <select class="form-select form-select-sm flex-grow-1" wicket:id="chapterSelect"></select>
+            <a class="btn btn-sm btn-outline-secondary px-2" wicket:id="nextChapter"
+               wicket:message="title:HelpContentPanel.nextChapter">
+                <i class="fa fa-arrow-right"></i>
+            </a>
+        </form>
+
+        <h5 class="mt-1 mb-3 fw-semibold" wicket:id="chapterTitle"></h5>
+
+        <div class="text-help lh-base overflow-y-auto overflow-x-hidden pe-1"
+             wicket:id="chapterContent"></div>
+    </div>
+</wicket:panel>
+</html>
