+++ gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/component/wizard/collapse/HelpContentPanel.html
@@ -8,34 +8,7 @@
 <html xmlns:wicket="http://wicket.apache.org">
 <wicket:panel>
     <div class="d-flex flex-column h-100 px-2">
-
-        <ul class="nav nav-tabs flex-nowrap" wicket:id="tabs">
-            <li class="nav-item" wicket:id="tab">
-                <a class="nav-link px-2 text-nowrap" wicket:id="tabLink">
-                    <i class="me-1" wicket:id="tabIcon"></i>
-                    <span wicket:id="tabLabel"></span>
-                </a>
-            </li>
-        </ul>
-
-        <form class="d-flex flex-row align-items-center gap-1 py-2" wicket:id="pagerForm">
-            <a class="btn btn-sm btn-outline-secondary px-2" wicket:id="previousChapter"
-               wicket:message="title:HelpContentPanel.previousChapter">
-                <i class="fa fa-arrow-left"></i>
-            </a>
-            <span class="border rounded px-2 py-1 small text-secondary text-nowrap"
-                  wicket:id="chapterCounter"></span>
-            <select class="form-select form-select-sm flex-grow-1" wicket:id="chapterSelect"></select>
-            <a class="btn btn-sm btn-outline-secondary px-2" wicket:id="nextChapter"
-               wicket:message="title:HelpContentPanel.nextChapter">
-                <i class="fa fa-arrow-right"></i>
-            </a>
-        </form>
-
-        <h5 class="mt-2 mb-3 fw-semibold" wicket:id="chapterTitle"></h5>
-
-        <div class="text-help lh-base overflow-y-auto overflow-x-hidden pe-1"
-             wicket:id="chapterContent"></div>
+        <div wicket:id="content"></div>
     </div>
 </wicket:panel>
 </html>
